+++ src/Options.ts
@@ -80,7 +80,13 @@ const options = {
   'storybook-address': {
     alias: 's',
     description:
-      'Storybook server address to test against instead of using a static build directory. If set, build-dir will be ignored.',
+      '** Deprecated! Use --storybook-url instead ** Storybook server address to test against instead of using a static build directory. If set, build-dir will be ignored.',
+    type: 'string' as const,
+  },
+  'storybook-url': {
+    alias: 'u',
+    description:
+      'Url to a running Storybook to test against. Alternative to --build-dir, which will be ignored if this is set.',
     type: 'string' as const,
   },
   timeout: {
@@ -107,7 +113,7 @@ export function parseOptions() {
     pattern: new RegExp(argv.pattern),
     reporter: argv.reporter,
     reporterOptions: getReporterOptions(argv['reporter-options']),
-    storybookAddress: argv.storybookAddress,
+    storybookUrl: argv.storybookAddress || argv.storybookUrl,
     timeout: argv.timeout,
   };
 }